requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2==2.9.10
drf-spectacular==0.27.2
python-dotenv==1.0.1
pytest==8.3.3
Faker==30.8.0
factory-boy==3.3.1
pytest-django==4.9.0
djangorestframework-simplejwt==5.3.1
django-debug-toolbar==4.4.6
django-autoslug==1.9.9
pillow==11.0.0
slugify==0.0.1
requests==2.32.3